+++ src/types/global.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-empty-interface */
 /* eslint-disable @typescript-eslint/no-namespace */
 
-import { IUser, UserRoleEnum } from '../types/user.type';
+import { IUser, UserRoleEnum } from './user.type';
 
 declare global {
 	namespace Express {
